+++ ldr_tools/src/ldraw.rs
@@ -623,6 +623,9 @@ impl Transform {
 mod tests {
     use super::*;
 
+    use glam::vec3;
+    use pretty_assertions::assert_eq;
+
     #[test]
     fn test_split_mpd_files() {
         let commands = vec![
@@ -632,10 +635,10 @@ mod tests {
             Command::SubFileRef(SubFileRefCmd {
                 color: 16,
                 transform: Transform {
-                    pos: Vec3::new(0.0, 0.0, 0.0),
-                    row0: Vec3::new(1.0, 0.0, 0.0),
-                    row1: Vec3::new(0.0, 1.0, 0.0),
-                    row2: Vec3::new(0.0, 0.0, 1.0),
+                    pos: vec3(0.0, 0.0, 0.0),
+                    row0: vec3(1.0, 0.0, 0.0),
+                    row1: vec3(0.0, 1.0, 0.0),
+                    row2: vec3(0.0, 0.0, 1.0),
                 },
                 file: "1.dat".to_string(),
             }),
@@ -646,10 +649,10 @@ mod tests {
             Command::SubFileRef(SubFileRefCmd {
                 color: 16,
                 transform: Transform {
-                    pos: Vec3::new(0.0, 0.0, 0.0),
-                    row0: Vec3::new(1.0, 0.0, 0.0),
-                    row1: Vec3::new(0.0, 1.0, 0.0),
-                    row2: Vec3::new(0.0, 0.0, 1.0),
+                    pos: vec3(0.0, 0.0, 0.0),
+                    row0: vec3(1.0, 0.0, 0.0),
+                    row1: vec3(0.0, 1.0, 0.0),
+                    row2: vec3(0.0, 0.0, 1.0),
                 },
                 file: "2.dat".to_string(),
             }),
@@ -684,10 +687,10 @@ mod tests {
             Command::SubFileRef(SubFileRefCmd {
                 color: 16,
                 transform: Transform {
-                    pos: Vec3::new(0.0, 0.0, 0.0),
-                    row0: Vec3::new(1.0, 0.0, 0.0),
-                    row1: Vec3::new(0.0, 1.0, 0.0),
-                    row2: Vec3::new(0.0, 0.0, 1.0),
+                    pos: vec3(0.0, 0.0, 0.0),
+                    row0: vec3(1.0, 0.0, 0.0),
+                    row1: vec3(0.0, 1.0, 0.0),
+                    row2: vec3(0.0, 0.0, 1.0),
                 },
                 file: "1.dat".to_string(),
             }),
@@ -697,10 +700,10 @@ mod tests {
             Command::SubFileRef(SubFileRefCmd {
                 color: 16,
                 transform: Transform {
-                    pos: Vec3::new(0.0, 0.0, 0.0),
-                    row0: Vec3::new(1.0, 0.0, 0.0),
-                    row1: Vec3::new(0.0, 1.0, 0.0),
-                    row2: Vec3::new(0.0, 0.0, 1.0),
+                    pos: vec3(0.0, 0.0, 0.0),
+                    row0: vec3(1.0, 0.0, 0.0),
+                    row1: vec3(0.0, 1.0, 0.0),
+                    row2: vec3(0.0, 0.0, 1.0),
                 },
                 file: "2.dat".to_string(),
             }),
@@ -764,8 +767,219 @@ mod tests {
         ";
 
         let commands = parse_raw(ldr_contents).unwrap();
-        // TODO: Check the actual commands.
-        assert_eq!(28, commands.len());
+        assert_eq!(
+            vec![
+                            Command::File(FileCmd {
+                                file: "main.ldr".to_string()
+                            }),
+                            Command::SubFileRef(SubFileRefCmd {
+                                color: 7,
+                                transform: Transform {
+                                    pos: vec3(0.0, 0.0, 0.0),
+                                    row0: vec3(1.0, 0.0, 0.0),
+                                    row1: vec3(0.0, 1.0, 0.0),
+                                    row2: vec3(0.0, 0.0, 1.0)
+                                },
+                                file: "819.dat".to_string()
+                            }),
+                            Command::SubFileRef(SubFileRefCmd {
+                                color: 4,
+                                transform: Transform {
+                                    pos: vec3(80.0, -8.0, 70.0),
+                                    row0: vec3(1.0, 0.0, 0.0),
+                                    row1: vec3(0.0, 1.0, 0.0),
+                                    row2: vec3(0.0, 0.0, 1.0)
+                                },
+                                file: "house.ldr".to_string()
+                            }),
+                            Command::SubFileRef(SubFileRefCmd {
+                                color: 4,
+                                transform: Transform {
+                                    pos: vec3(-70.0, -8.0, 20.0),
+                                    row0: vec3(0.0, 0.0, -1.0),
+                                    row1: vec3(0.0, 1.0, 0.0),
+                                    row2: vec3(1.0, 0.0, 0.0)
+                                },
+                                file: "house.ldr".to_string()
+                            }),
+                            Command::SubFileRef(SubFileRefCmd {
+                                color: 4,
+                                transform: Transform {
+                                    pos: vec3(50.0, -8.0, -20.0),
+                                    row0: vec3(0.0, 0.0, -1.0),
+                                    row1: vec3(0.0, 1.0, 0.0),
+                                    row2: vec3(1.0, 0.0, 0.0)
+                                },
+                                file: "house.ldr".to_string()
+                            }),
+                            Command::SubFileRef(SubFileRefCmd {
+                                color: 4,
+                                transform: Transform {
+                                    pos: vec3(0.0, -8.0, -30.0),
+                                    row0: vec3(1.0, 0.0, 0.0),
+                                    row1: vec3(0.0, 1.0, 0.0),
+                                    row2: vec3(0.0, 0.0, 1.0)
+                                },
+                                file: "house.ldr".to_string()
+                            }),
+                            Command::SubFileRef(SubFileRefCmd {
+                                color: 4,
+                                transform: Transform {
+                                    pos: vec3(-20.0, -8.0, 70.0),
+                                    row0: vec3(1.0, 0.0, 0.0),
+                                    row1: vec3(0.0, 1.0, 0.0),
+                                    row2: vec3(0.0, 0.0, 1.0)
+                                },
+                                file: "house.ldr".to_string()
+                            }),
+                            Command::File(FileCmd {
+                                file: "house.ldr".to_string()
+                            }),
+                            Command::SubFileRef(SubFileRefCmd {
+                                color: 16,
+                                transform: Transform {
+                                    pos: vec3(0.0, 0.0, 0.0),
+                                    row0: vec3(1.0, 0.0, 0.0),
+                                    row1: vec3(0.0, 1.0, 0.0),
+                                    row2: vec3(0.0, 0.0, 1.0)
+                                },
+                                file: "3023.dat".to_string()
+                            }),
+                            Command::SubFileRef(SubFileRefCmd {
+                                color: 16,
+                                transform: Transform {
+                                    pos: vec3(0.0, -24.0, 0.0),
+                                    row0: vec3(1.0, 0.0, 0.0),
+                                    row1: vec3(0.0, 1.0, 0.0),
+                                    row2: vec3(0.0, 0.0, 1.0)
+                                },
+                                file: "3065.dat".to_string()
+                            }),
+                            Command::SubFileRef(SubFileRefCmd {
+                                color: 16,
+                                transform: Transform {
+                                    pos: vec3(0.0, -48.0, 0.0),
+                                    row0: vec3(1.0, 0.0, 0.0),
+                                    row1: vec3(0.0, 1.0, 0.0),
+                                    row2: vec3(0.0, 0.0, 1.0)
+                                },
+                                file: "3065.dat".to_string()
+                            }),
+                            Command::SubFileRef(SubFileRefCmd {
+                                color: 16,
+                                transform: Transform {
+                                    pos: vec3(0.0, -72.0, 0.0),
+                                    row0: vec3(0.0, 0.0, -1.0),
+                                    row1: vec3(0.0, 1.0, 0.0),
+                                    row2: vec3(1.0, 0.0, 0.0)
+                                },
+                                file: "3044b.dat".to_string()
+                            }),
+                            Command::SubFileRef(SubFileRefCmd {
+                                color: 4,
+                                transform: Transform {
+                                    pos: vec3(0.0, -22.0, -10.0),
+                                    row0: vec3(1.0, 0.0, 0.0),
+                                    row1: vec3(0.0, 0.0, -1.0),
+                                    row2: vec3(0.0, 1.0, 0.0)
+                                },
+                                file: "sticker.ldr".to_string()
+                            }),
+                            Command::File(FileCmd {
+                                file: "sticker.ldr".to_string()
+                            }),
+                            Command::Comment(CommentCmd {
+                                text: "UNOFFICIAL PART".to_string()
+                            }),
+                            Command::Bfc(BfcCommand::Certify(Some(Winding::Ccw))),
+                            Command::SubFileRef(SubFileRefCmd {
+                                color: 16,
+                                transform: Transform {
+                                    pos: vec3(0.0, -0.25, 0.0),
+                                    row0: vec3(20.0, 0.0, 0.0),
+                                    row1: vec3(0.0, 0.25, 0.0),
+                                    row2: vec3(0.0, 0.0, 30.0)
+                                },
+                                file: "box5.dat".to_string()
+                            }),
+                            Command::Comment(
+                                CommentCmd {
+                                    text: "!TEXMAP START PLANAR   -20 -0.25 30   20 -0.25 30   -20 -0.25 -30   sticker.png".to_string()
+                                },
+                            ),
+                            Command::Quad(QuadCmd {
+                                color: 16,
+                                vertices: [
+                                    vec3(-20.0, -0.25, 30.0),
+                                    vec3(-20.0, -0.25, -30.0),
+                                    vec3(20.0, -0.25, -30.0),
+                                    vec3(20.0, -0.25, 30.0)
+                                ],
+                                uvs: None
+                            }),
+                            Command::Comment(CommentCmd {
+                                text: "!TEXMAP END".to_string()
+                            }),
+                            Command::Data(DataCmd {
+                                file: "sticker.png".to_string()
+                            }),
+                            Command::Base64Data(Base64DataCmd {
+                                data: vec![
+                                    137, 80, 78, 71, 13, 10, 26, 10, 0, 0, 0, 13, 73, 72, 68, 82, 0, 0, 0, 80,
+                                    0, 0, 0, 120, 8, 2, 0, 0, 0, 234, 140, 226, 161, 0, 0, 0, 1, 115, 82, 71,
+                                    66, 0, 174, 206, 28, 233, 0, 0, 0, 4, 103, 65, 77, 65, 0, 0, 177
+                                ]
+                            }),
+                            Command::Base64Data(Base64DataCmd {
+                                data: vec![
+                                    143, 11, 252, 97, 5, 0, 0, 0, 9, 112, 72, 89, 115, 0, 0, 14, 195, 0, 0, 14,
+                                    195, 1, 199, 111, 168, 100, 0, 0, 1, 20, 73, 68, 65, 84, 120, 94, 237, 219,
+                                    189, 13, 194, 48, 20, 0, 97, 147, 149, 160, 64, 180, 172, 64, 205, 20, 140,
+                                    193, 20
+                                ]
+                            }),
+                            Command::Base64Data(Base64DataCmd {
+                                data: vec![
+                                    212, 20, 44, 64, 139, 40, 96, 11, 122, 54, 64, 24, 197, 5, 127, 37, 14,
+                                    214, 189, 251, 138, 144, 84, 209, 233, 73, 142, 20, 156, 209, 245, 178, 75,
+                                    145, 116, 229, 55, 12, 131, 233, 12, 166, 51, 152, 238, 229, 57, 124, 155,
+                                    174, 202, 25, 75, 119, 92, 151, 51
+                                ]
+                            }),
+                            Command::Base64Data(Base64DataCmd {
+                                data: vec![
+                                    39, 252, 152, 240, 118, 127, 232, 47, 1, 22, 243, 89, 62, 134, 158, 176,
+                                    193, 116, 6, 211, 25, 76, 103, 48, 157, 193, 116, 6, 211, 25, 76, 103, 48,
+                                    157, 193, 116, 6, 211, 25, 76, 103, 48, 157, 255, 45, 209, 185, 139, 135,
+                                    206, 96, 58, 131, 233, 12, 166, 51
+                                ]
+                            }),
+                            Command::Base64Data(Base64DataCmd {
+                                data: vec![
+                                    152, 206, 96, 58, 131, 233, 12, 166, 51, 152, 206, 96, 186, 112, 193, 127,
+                                    123, 77, 251, 252, 234, 52, 171, 119, 235, 183, 27, 57, 225, 148, 78, 203,
+                                    113, 127, 89, 201, 100, 115, 206, 199, 175, 19, 254, 237, 215, 7, 159, 251,
+                                    255, 51, 23, 45, 58, 131, 233, 12, 166
+                                ]
+                            }),
+                            Command::Base64Data(Base64DataCmd {
+                                data: vec![
+                                    51, 152, 206, 96, 58, 131, 233, 12, 166, 51, 152, 206, 96, 58, 131, 233,
+                                    12, 166, 51, 152, 206, 96, 58, 131, 233, 12, 166, 51, 152, 206, 96, 186,
+                                    112, 193, 109, 237, 241, 168, 193, 93, 60, 109, 76, 120, 48, 46, 90, 116,
+                                    6, 211, 25, 76, 103, 48, 157, 193, 116
+                                ]
+                            }),
+                            Command::Base64Data(Base64DataCmd {
+                                data: vec![
+                                    6, 211, 25, 76, 215, 202, 103, 60, 131, 137, 61, 225, 8, 92, 180, 232, 130,
+                                    5, 167, 116, 7, 252, 32, 76, 135, 174, 154, 7, 46, 0, 0, 0, 0, 73, 69, 78,
+                                    68, 174, 66, 96, 130
+                                ]
+                            })
+                        ],
+            commands
+        );
     }
 
     #[test]
@@ -773,7 +987,7 @@ mod tests {
         let cmd0 = Command::Comment(CommentCmd::new("this is a comment"));
         let cmd1 = Command::Line(LineCmd {
             color: 16,
-            vertices: [Vec3::new(0.0, 0.0, 0.0), Vec3::new(1.0, 1.0, 1.0)],
+            vertices: [vec3(0.0, 0.0, 0.0), vec3(1.0, 1.0, 1.0)],
         });
         assert_eq!(
             parse_raw(b"0 this is a comment\n2 16 0 0 0 1 1 1").unwrap(),
@@ -784,10 +998,10 @@ mod tests {
         let cmd1 = Command::SubFileRef(SubFileRefCmd {
             color: 16,
             transform: Transform {
-                pos: Vec3::new(0.0, 0.0, 0.0),
-                row0: Vec3::new(1.0, 0.0, 0.0),
-                row1: Vec3::new(0.0, 1.0, 0.0),
-                row2: Vec3::new(0.0, 0.0, 1.0),
+                pos: vec3(0.0, 0.0, 0.0),
+                row0: vec3(1.0, 0.0, 0.0),
+                row1: vec3(0.0, 1.0, 0.0),
+                row2: vec3(0.0, 0.0, 1.0),
             },
             file: "aa/aaaaddd".to_string(),
         });
@@ -801,10 +1015,10 @@ mod tests {
         let cmd1 = Command::SubFileRef(SubFileRefCmd {
             color: 16,
             transform: Transform {
-                pos: Vec3::new(0.0, 0.0, 0.0),
-                row0: Vec3::new(1.0, 0.0, 0.0),
-                row1: Vec3::new(0.0, 1.0, 0.0),
-                row2: Vec3::new(0.0, 0.0, 1.0),
+                pos: vec3(0.0, 0.0, 0.0),
+                row0: vec3(1.0, 0.0, 0.0),
+                row1: vec3(0.0, 1.0, 0.0),
+                row2: vec3(0.0, 0.0, 1.0),
             },
             file: "aa/aaaaddd".to_string(),
         });
@@ -819,20 +1033,20 @@ mod tests {
         let cmd0 = Command::SubFileRef(SubFileRefCmd {
             color: 16,
             transform: Transform {
-                pos: Vec3::new(0.0, 0.0, 0.0),
-                row0: Vec3::new(1.0, 0.0, 0.0),
-                row1: Vec3::new(0.0, 1.0, 0.0),
-                row2: Vec3::new(0.0, 0.0, 1.0),
+                pos: vec3(0.0, 0.0, 0.0),
+                row0: vec3(1.0, 0.0, 0.0),
+                row1: vec3(0.0, 1.0, 0.0),
+                row2: vec3(0.0, 0.0, 1.0),
             },
             file: "aa/aaaaddd".to_string(),
         });
         let cmd1 = Command::SubFileRef(SubFileRefCmd {
             color: 16,
             transform: Transform {
-                pos: Vec3::new(0.0, 0.0, 0.0),
-                row0: Vec3::new(1.0, 0.0, 0.0),
-                row1: Vec3::new(0.0, 1.0, 0.0),
-                row2: Vec3::new(0.0, 0.0, 1.0),
+                pos: vec3(0.0, 0.0, 0.0),
+                row0: vec3(1.0, 0.0, 0.0),
+                row1: vec3(0.0, 1.0, 0.0),
+                row2: vec3(0.0, 0.0, 1.0),
             },
             file: "aa/aaaaddd".to_string(),
         });