+++ CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## unreleased
+### Fixed
+* Fixed an issue where log messages would not appear in the Blender console.
+
 ## 0.4.10 - 2025-07-11
 ### Fixed
 * Fixed an issue where the incorrect texture would be assigned to a face in some cases.